back-end/ungleich-otp-master/ungleichotpserver/requirements.txt
pyotp>=2.2.6
django>=2.1.2
djangorestframework

# Recommended
markdown
django-filter
